requirements.txt
aind-data-schema==1.2.0
aind-data-schema-models==0.7.5
aind-metadata-mapper==0.21.0
aind_metadata_service==0.20.0
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.2.0
beautifulsoup4==4.13.3
bidict==0.23.1
build==1.2.2.post1
build-metadata==0.2
cachetools==5.5.2
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
CoverGenerator==0.2.0
cssselect2==0.8.0
Cython==3.0.12
defusedxml==0.7.1
dictdiffer==0.9.0
distro==1.9.0
dj-stripe==2.9.0
Django==5.1.7
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-environ==0.12.0
django-freeze==0.11.1
django-metadata==0.2.0
django-pghistory==3.5.4
django-pgtrigger==4.13.3
django-sslserver==0.22
djangorestframework==3.15.2
dnspython==2.7.0
docker==2.7.0
docker-pycreds==0.4.0
docutils==0.21.2
drf-spectacular==0.28.0
dynamic-metadata==0.1.0
email_validator==2.2.0
fastapi==0.115.11
fastopenapi==0.2.1
for-django-projects==1.3.3
gitlab-api==1.0.76
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
HTML-String-Tools==0.3.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
id==1.5.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
lazy_loader==0.4
legacy-cgi==2.6.2
libxml2dom==0.5
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
Metadata-Magic==0.12.10
metadata_parser==0.12.2
more-itertools==10.6.0
nh3==0.2.21
numpy==2.2.3
openai==1.66.3
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==6.1.1
psycopg==3.2.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pyproject-metadata==0.9.1
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-metadata==3.1.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.20
Python-Print-Tools==0.2.2
python-socketio==5.12.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
readme_renderer==44.0
redis==5.2.1
referencing==0.36.2
reflex==0.7.2
reflex-hosting-cli==0.1.35
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
scanimage-tiff-reader==1.4.1.4
semver==3.0.4
setup-py-upgrade==1.3.1
setuptools==76.0.0
setuptools-metadata==0.1.5
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sql_metadata==2.15.0
SQLAlchemy==2.0.39
sqlmodel==0.0.24
sqlparse==0.5.3
starlette==0.46.1
starlette-admin==0.14.1
stripe==11.6.0
tabulate==0.9.0
tenacity==9.0.0
tinycss2==1.4.0
tomlkit==0.13.2
tqdm==4.67.1
twine==6.1.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
unixpackage==0.4.7
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
whitenoise==6.9.0
wrapt==1.17.2
wsproto==1.2.0
xmltodict==0.14.2
z3c.recipe.staticlxml==0.10
zc.buildout==4.1.9
zc.recipe.cmmi==4.0
zc.recipe.egg==3.0.0
zipp==3.21.0
